Add main game component
